rename test db file
